requirements.txt
requests
pandas
psycopg2-binary
boto3
scipy
numpy